+++ sheriff/CHANGELOG.md
@@ -1,5 +1,24 @@
-# Revision history for code-checker
+# Revision history for sheriff
+
+## 0.2.0.4
+* Fix modules names not getting matched due to wildcard support logical bug
+
+## 0.2.0.3
+* Fix non-exhaustive pattern
+* Fix wildcard character matching for asterisk operator
+* Customizable asterisk matching based on use case
+
+## 0.2.0.2
+* Remove support for wilcard character "*" in function rules (to be fixed)
+
+## 0.2.0.1
+* Add derived signature from the arg types for Signature Check rules
+
+## 0.2.0.0
+* Add allowed modules list for function rule
+* Add support for wilcard character "*" in function rules
+* Add support for blocking functions with a particular type signature (Signature only)
+* Add test rules in sheriff rules
 
 ## 0.1.0.0 -- 2024-03-20
-
 * First version. Basic rules based compilation error.